+++ src/utilities.h
@@ -46,22 +46,22 @@ bool fileExists(const string &path);
 bool dirExists(const string &path);
 bool rmtree(string path);
 
-int max (int a, int b);
-int min (int a, int b);
+int max(int a, int b);
+int min(int a, int b);
 int constrain(int x, int imin, int imax);
 
-int evalIntConf (int val, int def, int imin, int imax);
-int evalIntConf (int *val, int def, int imin, int imax);
-const string &evalStrConf (const string &val, const string &def);
-const string &evalStrConf (string *val, const string &def);
+int evalIntConf(int val, int def, int imin, int imax);
+int evalIntConf(int *val, int def, int imin, int imax);
+const string &evalStrConf(const string &val, const string &def);
+const string &evalStrConf(string *val, const string &def);
 
-float max (float a, float b);
-float min (float a, float b);
-float constrain (float x, float imin, float imax);
+float max(float a, float b);
+float min(float a, float b);
+float constrain(float x, float imin, float imax);
 
-bool split (vector<string> &vec, const string &str, const string &delim, bool destructive=true);
+bool split(vector<string> &vec, const string &str, const string &delim, bool destructive=true);
 
-int intTransition(int from, int to, long int tickStart, long duration=500, long tickNow=-1);
+int intTransition(int from, int to, int32_t tickStart, int32_t duration = 500, int32_t tickNow = -1);
 
 string exec(const char* cmd);
 string real_path(const string &path);
